+++ packages/opentelemetry-node/test/instr-http.test.js
@@ -6,7 +6,7 @@
 // Test that 'http' instrumentation generates the telemetry we expect.
 
 const test = require('tape');
-const {runTestFixtures, filterOutGcpDetectorSpans} = require('./testutils');
+const {runTestFixtures} = require('./testutils');
 
 /** @type {import('./testutils').TestFixture[]} */
 const testFixtures = [
@@ -19,7 +19,7 @@ const testFixtures = [
         },
         // verbose: true,
         checkTelemetry: (t, col) => {
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 1);
             const span = spans[0];
             t.equal(span.scope.name, '@opentelemetry/instrumentation-http');
@@ -39,7 +39,7 @@ const testFixtures = [
         },
         // verbose: true,
         checkTelemetry: (t, col) => {
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 1);
             const span = spans[0];
             t.equal(span.scope.name, '@opentelemetry/instrumentation-http');
@@ -58,7 +58,7 @@ const testFixtures = [
         },
         // verbose: true,
         checkTelemetry: (t, col) => {
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 1);
             const span = spans[0];
             t.equal(span.scope.name, '@opentelemetry/instrumentation-http');
@@ -78,7 +78,7 @@ const testFixtures = [
         },
         // verbose: true,
         checkTelemetry: (t, col) => {
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 1);
             const span = spans[0];
             t.equal(span.scope.name, '@opentelemetry/instrumentation-http');
@@ -98,7 +98,7 @@ const testFixtures = [
         },
         // verbose: true,
         checkTelemetry: (t, col) => {
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 1);
             const span = spans[0];
             t.equal(span.scope.name, '@opentelemetry/instrumentation-http');
@@ -125,7 +125,7 @@ const testFixtures = [
             //     ------ trace b8467d (2 spans) ------
             //            span bc8a2c "POST" (3.8ms, SPAN_KIND_CLIENT, POST http://127.0.0.1:64972/echo -> 200)
             //      +2ms `- span 4e7adf "POST" (1.1ms, SPAN_KIND_SERVER, POST http://127.0.0.1:64972/echo -> 200)
-            const spans = filterOutGcpDetectorSpans(col.sortedSpans);
+            const spans = col.sortedSpans;
             t.equal(spans.length, 4);
             t.equal(spans[0].scope.name, '@opentelemetry/instrumentation-http');
             t.equal(spans[0].name, 'GET');